fix: serve GitHub e2e helper as JavaScript
